mysite/requirements.txt
asgiref==3.8.1
construct==2.5.3
Django==5.1.3
django-taggit==6.1.0
haystack==0.42
Markdown==3.7
pefile==2024.8.26
python-ptrace==0.9.9
six==1.16.0
sqlparse==0.5.2
